Rename backend to pyro.ops.einsum.deferred
